as.IFAgroup {rpf}	R Documentation

Convert an OpenMx MxModel object into an IFA group

Description

When “minItemsPerScore” is passed, EAP scores will be computed from the data and stored. Scores are required for some diagnostic tests. See discussion of “minItemsPerScore” in EAPscores.

Usage

as.IFAgroup(mxModel, data = NULL, container = NULL, ...,
  minItemsPerScore = NULL)

Arguments

mxModel
MxModel object

data
observed data (otherwise the data will be taken from the mxModel)

container
an MxModel in which to search for the latent distribution matrices

...
Not used. Forces remaining arguments to be specified by name.

minItemsPerScore
minimum number of items required to compute a score (also see description)

Value

a groups with item parameters and latent distribution

See Also

ifaTools

[Package rpf version 0.56.1 Index]