bestToOmit {rpf}	R Documentation

Identify the columns with most missing data

Description

If a reference column is given then only rows that are not missing on the reference column are considered. Otherwise all rows are considered.

Usage

bestToOmit(grp, omit, ref = NULL)

Arguments

grp
an IFA group

omit
the maximum number of items to omit

ref
the reference column (optional)

[Package rpf version 0.56.1 Index]