compressDataFrame {rpf}	R Documentation

Compress a data frame into unique rows and frequencies

Description

Compress a data frame into unique rows and frequency counts.

Usage

compressDataFrame(tabdata, freqColName = "freq", .asNumeric = FALSE)

Arguments

tabdata
An object of class data.frame

freqColName
Column name to contain the frequencies

.asNumeric
logical. Whether to cast the frequencies to the numeric type

Value

Returns a compressed data frame

Examples

df <- as.data.frame(matrix(c(sample.int(2, 30, replace=TRUE)), 10, 3))
compressDataFrame(df)

[Package rpf version 1.0.3 Index]