expandDataFrame {rpf}	R Documentation

Expand summary table of patterns and frequencies

Description

Expand a summary table of unique response patterns to a full sized data-set.

Usage

expandDataFrame(tabdata, freqName = NULL)

Arguments

tabdata
An object of class data.frame with the unique response patterns and the number of frequencies

freqName
Column name containing the frequencies

Value

Returns a data frame with all the response patterns

Author(s)

Based on code by Phil Chalmers rphilip.chalmers@gmail.com

Examples

data(LSAT7)
expandDataFrame(LSAT7, freqName="freq")

[Package rpf version 1.0.3 Index]