fromFactorLoading {rpf}	R Documentation

Convert factor loadings to response function slopes

Description

Convert factor loadings to response function slopes

Usage

fromFactorLoading(loading, ogive = rpf.ogive)

Arguments

loading
a matrix with items in the rows and factors in the columns

ogive
the ogive constant (default rpf.ogive)

Value

a slope matrix with items in the columns and factors in the rows

See Also

Other factor model equivalence: fromFactorThreshold(), toFactorLoading(), toFactorThreshold()

[Package rpf version 1.0.15 Index]