fromFactorThreshold {rpf}	R Documentation

Convert factor thresholds to response function intercepts

Description

Convert factor thresholds to response function intercepts

Usage

fromFactorThreshold(threshold, loading, ogive = rpf.ogive)

Arguments

threshold
a matrix with items in the columns and thresholds in the rows

loading
a matrix with items in the rows and factors in the columns

ogive
the ogive constant (default rpf.ogive)

Value

an item intercept matrix with items in the columns and intercepts in the rows

[Package rpf version 0.56.1 Index]