rpf.dLL {rpf}	R Documentation

Item parameter derivatives

Description

Evaluate the partial derivatives of the log likelihood with respect to each parameter at where with weight.

Usage

rpf.dLL(m, param, where, weight)

Arguments

m
item model

param
item parameters

where
location in the latent space

weight
per outcome weights (typically derived by observation)

Details

It is not easy to write an example for this function. To evaluate the derivative, you need to sum the derivatives across a quadrature. You also need response outcome weights at each quadrature point. It is not anticipated that this function will be often used in R code. It's mainly to expose a C-level function for occasional debugging.

Value

first and second order partial derivatives of the log likelihood evaluated at where. For p parameters, the first p values are the first derivative and the next p(p+1)/2 columns are the lower triangle of the second derivative.

See Also

The numDeriv package.

[Package rpf version 1.0.3 Index]