rpf.dTheta {rpf}	R Documentation

Item derivatives with respect to the location in the latent space

Description

Evaluate the partial derivatives of the response probability with respect to ability. See rpf.info for an application.

Usage

rpf.dTheta(m, param, where, dir)

Arguments

m
item model

param
item parameters

where
location in the latent distribution

dir
if more than 1 factor, a basis vector

[Package rpf version 1.0.4 Index]