rpf.drm {rpf}	R Documentation

Create a dichotomous response model

Description

For slope vector a, intercept c, pseudo-guessing parameter g, upper bound u, and latent ability vector theta, the response probability function is

P(pick=0|a,c,g,u,th) = 1-P(pick=1|a,c,g,u,th)

P(pick=1|a,c,g,u,th) = g+(u-g)/(1+exp(-(a th + c)))

Usage

rpf.drm(factors = 1, multidimensional = TRUE, poor = FALSE)

Arguments

factors
the number of factors

multidimensional
whether to use a multidimensional model. Defaults to TRUE.

poor
if TRUE, use the traditional parameterization of the 1d model instead of the slope-intercept parameterization

Details

The pseudo-guessing and upper bound parameter are specified in logit units (see logit).

For discussion on the choice of priors see Cai, Yang, and Hansen (2011, p. 246).

Value

an item model

References

Cai, L., Yang, J. S., & Hansen, M. (2011). Generalized Full-Information Item Bifactor Analysis. Psychological Methods, 16(3), 221-248.

See Also

Other response model: rpf.gpcmp(), rpf.grmp(), rpf.grm(), rpf.lmp(), rpf.mcm(), rpf.nrm()

Examples

spec <- rpf.drm()
rpf.prob(spec, rpf.rparam(spec), 0)

[Package rpf version 1.0.3 Index]