rpf.grm {rpf}	R Documentation

Create a graded response model

Description

For outcomes k in 0 to K, slope vector a, intercept vector c, and latent ability vector theta, the response probability function is

P(pick=0|a,c,th) = 1-P(pick=1|a,c_1,th)

P(pick=k|a,c,th) = 1/(1+exp(-(a th + c_k))) - 1/(1+exp(-(a th + c_(k+1))))

P(pick=K|a,c,th) = 1/(1+exp(-(a th + c_K)))

Usage

rpf.grm(outcomes = 2, factors = 1, multidimensional = TRUE)

Arguments

outcomes
The number of choices available

factors
the number of factors

multidimensional
whether to use a multidimensional model. Defaults to TRUE.

Details

The graded response model was designed for a item with a series of dependent parts where a higher score implies that easier parts of the item were surmounted. If there is any chance your polytomous item has independent parts then consider rpf.nrm. If your categories cannot cross then the graded response model provides a little more information than the nominal model. Stronger a priori assumptions offer provide more power at the cost of flexibility.

Value

an item model

See Also

Other response model: rpf.drm(), rpf.gpcmp(), rpf.grmp(), rpf.lmp(), rpf.mcm(), rpf.nrm()

Examples

spec <- rpf.grm()
rpf.prob(spec, rpf.rparam(spec), 0)

[Package rpf version 1.0.4 Index]