rpf.grmp {rpf}	R Documentation

Create monotonic polynomial graded response (GR-MP) model

Description

The GR-MP model replaces the linear predictor of the graded response model (Samejima, 1969, 1972) with a monotonic polynomial (Falk, conditionally accepted).

Usage

rpf.grmp(outcomes = 2, q = 0, multidimensional = FALSE)

Arguments

outcomes
The number of possible response categories. When equal to 2, the model reduces to the logistic function of a monotonic polynomial (LMP).

q
a non-negative integer that controls the order of the polynomial (2q+1) with a default of q=0 (1st order polynomial = graded response model).

multidimensional
whether to use a multidimensional model. Defaults to FALSE. The multidimensional version is not yet available.

Details

Given its relationship to the graded response model, the GR-MP is constructed in an analogous way:

P(pick=0|a,c,theta) = 1-1/(1+exp(-(xi_1 + m(theta;lambda,alpha,tau)))

P(pick=k|a,c,theta) = 1/(1+exp(-(xi_k + m(theta;lambda,alpha,tau))) - 1/(1+exp(-(xi_(k+1) + m(theta;lambda,alpha,tau))))

P(pick=K|λ,α,τ,theta) = 1/(1+exp(-(xi_K + m(theta;lambda,alpha,tau))))

The order of the polynomial is always odd and is controlled by the user specified non-negative integer, q. The model contains 1+(outcomtes-1)+2*q parameters and are used as input to the rpf.prob or rpf.dTheta functions in the following order: lambda - slope of the item model when q=0, xi - a (outcomes-1)-length vector of intercept parameters, alpha and tau - two parameters that control bends in the polynomial. These latter parameters are repeated in the same order for models with q>0. For example, a q=2 polynomial with 3 categories will have an item parameter vector of: lambda, xi1, xi2, alpha1, tau1, alpha2, tau2.

As with other monotonic polynomial-based item models (e.g., rpf.lmp), the polynomial looks like the following:

m(theta) = b_1*theta + b_2*theta^2 + … + b_(2q+1)*theta^{2q+1}

However, the coefficients, b, are not directly estimated, but are a function of the item parameters, and the parameterization of the GR-MP is different than that currently appearing for the logistic function of a monotonic polynomial (LMP; rpf.lmp) and monotonic polynomial generalized partial credit (GPC-MP; rpf.gpcmp) models. In particular, the polynomial is parameterized such that boundary descrimination functions for the GR-MP will be all monotonically increasing or decreasing for any given item. This allows the possibility of items that load either negatively or positively on the latent trait, as is common with reverse-worded items in non-cognitive tests (e.g., personality).

The derivative m'(theta;lambda,alpha,tau) is parameterized in the following way:

m'(theta) = m'(theta;lambda,alpha,tau) = lambda ∏_{u=1}^q (1-2*alpha_u*theta + (alpha_u^2 + exp(tau_u))*theta^2) (if q > 0) \ lambda (if q = 0)

Note that the only difference between the GR-MP and these other models is that lambda is not re-parameterized and may take on negative values. When lambda is negative, it is analogous to having a negative loading or a monotonically decreasing function.

Value

an item model

References

Falk, C. F. (conditionally accepted). The monotonic polynomial graded response model: Implementation and a comparative study. Applied Psychological Measurement.

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores. Psychometric Monographs, 17.

Samejima, F. (1972). A general model of free-response data. Psychometric Monographs, 18.

See Also

Other response model: rpf.drm(), rpf.gpcmp(), rpf.grm(), rpf.lmp(), rpf.mcm(), rpf.nrm()

Examples

spec <- rpf.grmp(5,2) # 5-category, 3rd order polynomial
theta<-seq(-3,3,.1)
p<-rpf.prob(spec, c(2.77,2,1,0,-1,.89,-8.7,-.74,-8.99),theta)

[Package rpf version 1.0.4 Index]