rpf.id_of {rpf}	R Documentation

Convert an rpf item model name to an ID

Description

This is an internal function and should not be used.

Usage

rpf.id_of(name)

Arguments

name
name of the item model (string)

Value

the integer ID assigned to the given model

[Package rpf version 1.0.4 Index]