rpf.info {rpf}	R Documentation

Map an item model, item parameters, and person trait score into a information vector

Description

Map an item model, item parameters, and person trait score into a information vector

Usage

rpf.info(ii, ii.p, where, basis = 1)

Arguments

ii
an item model

ii.p
item parameters

where
the location in the latent distribution

basis
if more than 1 factor, a positive basis vector

Value

Fisher information

References

Dodd, B. G., De Ayala, R. J. & Koch, W. R. (1995). Computerized adaptive testing with polytomous items. Applied psychological measurement 19(1), 5-22.

Examples

i1 <- rpf.drm()
i1.p <- c(.6,1,.1,.95)
theta <- seq(0,3,.05)
plot(theta, rpf.info(i1, i1.p, t(theta)), type="l")

[Package rpf version 0.56.1 Index]