Class rpf.mdim.nrm {rpf}	R Documentation

The nominal response item model (both unidimensional and multidimensional models have the same parameterization).

Description

The nominal response item model (both unidimensional and multidimensional models have the same parameterization).

[Package rpf version 1.0.4 Index]