rpf.mean.info {rpf}	R Documentation

Find the point where an item provides mean maximum information

Description

This is a point estimate of the mean difficulty of items that do not offer easily interpretable parameters such as the Generalized PCM. Since the information curve may not be unimodal, this function integrates across the latent space.

Usage

rpf.mean.info(spec, param, grain = 0.1)

Arguments

spec
list of item specs

param
list or matrix of item parameters

grain
the step size for numerical integration (optional)

[Package rpf version 1.0.4 Index]