rpf.modify {rpf}	R Documentation

Create a similar item specification with the given number of factors

Description

Create a similar item specification with the given number of factors

Usage

rpf.modify(m, factors)

Arguments

m
item model

factors
the number of factors/dimensions

Examples

s1 <- rpf.grm(factors=3)
rpf.rparam(s1)
s2 <- rpf.modify(s1, 1)
rpf.rparam(s2)

[Package rpf version 1.0.3 Index]