rpf.nrm {rpf}	R Documentation

Create a nominal response model

Description

This function instantiates a nominal response model.

Usage

rpf.nrm(outcomes = 3, factors = 1, T.a = "trend", T.c = "trend")

Arguments

outcomes
The number of choices available

factors
the number of factors

T.a
the T matrix for slope parameters

T.c
the T matrix for intercept parameters

Details

The transformation matrices T.a and T.c are chosen by the analyst and not estimated. The T matrices must be invertible square matrices of size outcomes-1. As a shortcut, either T matrix can be specified as "trend" for a Fourier basis or as "id" for an identity basis. The response probability function is

a = T_a α

c = T_c γ

\mathrm P(\mathrm{pick}=k|s,a_k,c_k,θ) = C\ \frac{1}{1+\exp(-(s θ a_k + c_k))}

where a_k and c_k are the result of multiplying two vectors of free parameters α and γ by fixed matrices T_a and T_c, respectively; a_0 and c_0 are fixed to 0 for identification; and C is a normalizing factor to ensure that ∑_k \mathrm P(\mathrm{pick}=k) = 1.

Value

an item model

References

Thissen, D., Cai, L., & Bock, R. D. (2010). The Nominal Categories Item Response Model. In M. L. Nering & R. Ostini (Eds.), Handbook of Polytomous Item Response Theory Models (pp. 43–75). Routledge.

See Also

Other response model: rpf.drm(), rpf.gpcmp(), rpf.grmp(), rpf.grm(), rpf.lmp(), rpf.mcm()

Examples

spec <- rpf.nrm()
rpf.prob(spec, rpf.rparam(spec), 0)
# typical parameterization for the Generalized Partial Credit Model
gpcm <- function(outcomes) rpf.nrm(outcomes, T.c=lower.tri(diag(outcomes-1),TRUE) * -1)
spec <- gpcm(4)
rpf.prob(spec, rpf.rparam(spec), 0)

[Package rpf version 1.0.3 Index]