rpf.rparam {rpf}	R Documentation

Generates item parameters

Description

This function generates random item parameters. The version argument is available if you are writing a test that depends on reproducable random parameters (using set.seed).

Usage

rpf.rparam(m, version = 2L)

Arguments

m
an item model

version
the version of random parameters

Value

item parameters

Examples

i1 <- rpf.drm()
rpf.rparam(i1)

[Package rpf version 0.56.1 Index]