sumScoreEAPTest {rpf}	R Documentation

Conduct the sum-score EAP distribution test

Description

Conduct the sum-score EAP distribution test

Usage

sumScoreEAPTest(grp, ..., qwidth = 6, qpoints = 49L, .twotier = TRUE)

Arguments

grp
a list with spec, param, mean, and cov

...
Not used. Forces remaining arguments to be specified by name.

qwidth
positive width of quadrature in Z units

qpoints
number of quadrature points

.twotier
whether to enable the two-tier optimization

References

Li, Z., & Cai, L. (2012, July). Summed score likelihood based indices for testing latent variable distribution fit in Item Response Theory. Paper presented at the annual International Meeting of the Psychometric Society, Lincoln, NE. Retrieved from http://www.cse.ucla.edu/downloads/files/SD2-final-4.pdf

[Package rpf version 0.56.1 Index]