tabulateRows {rpf}	R Documentation

Tabulate data.frame rows

Description

Like tabulate but entire rows are the unit of tabulation. The data.frame is not sorted, but must be sorted already.

Usage

tabulateRows(observed)

Arguments

observed
a sorted data.frame holding ordered factors in every column

See Also

orderCompletely

Examples

df <- as.data.frame(matrix(c(sample.int(2, 30, replace=TRUE)), 10, 3))
df <- df[orderCompletely(df),]
tabulateRows(df)

[Package rpf version 1.0.4 Index]