toFactorLoading {rpf}	R Documentation

Convert response function slopes to factor loadings

Description

All slopes are divided by the ogive constant. Then the following transformation is applied to the slope matrix,

Usage

toFactorLoading(slope, ogive = rpf.ogive)

Arguments

slope
a matrix with items in the columns and slopes in the rows

ogive
the ogive constant (default rpf.ogive)

Details

\frac{\mathrm{slope}}{≤ft[ 1 + \mathrm{rowSums}(\mathrm{slope}^2) \right]^\frac{1}{2}}

Value

a factor loading matrix with items in the rows and factors in the columns

See Also

Other factor model equivalence: fromFactorLoading(), fromFactorThreshold(), toFactorThreshold()

[Package rpf version 1.0.3 Index]