toFactorThreshold {rpf}	R Documentation

Convert response function intercepts to factor thresholds

Description

Convert response function intercepts to factor thresholds

Usage

toFactorThreshold(intercept, slope, ogive = rpf.ogive)

Arguments

intercept
a matrix with items in the columns and intercepts in the rows

slope
a matrix with items in the columns and slopes in the rows

ogive
the ogive constant (default rpf.ogive)

Value

a factor threshold matrix with items in the columns and factor thresholds in the rows

See Also

Other factor model equivalence: fromFactorLoading(), fromFactorThreshold(), toFactorLoading()

[Package rpf version 1.0.3 Index]