write.flexmirt {rpf}	R Documentation

Write a flexMIRT PRM file

Description

Formats item parameters in the way that flexMIRT expects to read them. Use read.flexmirt to see what shape the groups parameter of this function should take.

Usage

write.flexmirt(groups, file = NULL, fileEncoding = "")

Arguments

groups
a list of groups each with items and latent parameters

file
the destination file name

fileEncoding
how to encode the text file (optional)

Details

NOTE: Support for the graded response model may not be complete.

[Package rpf version 0.56.1 Index]