SitemFit {rpf}	R Documentation

Compute the S fit statistic for a set of items

Description

Runs SitemFit1 for every item and accumulates the results.

Usage

SitemFit(grp, ..., method = "pearson", log = TRUE, qwidth = 6,
  qpoints = 49L, alt = FALSE, omit = 0L, .twotier = TRUE,
  .parallel = TRUE)

Arguments

grp
a list with spec, param, mean, cov, data, and the free variable pattern

...
Not used. Forces remaining arguments to be specified by name.

method
whether to use a pearson or rms test

log
whether to return pvalues in log units

qwidth
the positive width of the quadrature in Z units

qpoints
the number of quadrature points

alt
whether to include the item of interest in the denominator

omit
number of items to omit (a single number) or a list of the length the number of items

.twotier
whether to enable the two-tier optimization

.parallel
whether to take advantage of multiple CPUs (default TRUE)

Value

a list of output from SitemFit1

Examples

grp <- list(spec=list())
grp$spec[1:20] <- rpf.grm()
grp$param <- sapply(grp$spec, rpf.rparam)
colnames(grp$param) <- paste("i", 1:20, sep="")
grp$mean <- 0
grp$cov <- diag(1)
grp$free <- grp$param != 0
grp$data <- rpf.sample(500, grp=grp)
SitemFit(grp)

[Package rpf version 0.56.1 Index]