fromFactorLoading {rpf}	R Documentation

Convert factor loadings to response function slopes

Description

Convert factor loadings to response function slopes

Usage

fromFactorLoading(loading, ogive = rpf.ogive)

Arguments

loading
a matrix with items in the rows and factors in the columns

ogive
the ogive constant (default rpf.ogive)

Value

a slope matrix with items in the columns and factors in the rows

[Package rpf version 0.56.1 Index]